+++ core/encoding/json/marshal.odin
@@ -190,13 +190,13 @@ marshal_to_writer :: proc(w: io.Writer, v: any, opt: ^Marshal_Options) -> (err: 
         if opt.write_uint_as_hex && (opt.spec == .JSON5 || opt.spec == .MJSON) {
             switch i in a {
             case u8, u16, u32, u64, u128:
-                s = strconv.write_bits_128(buf[:], u, 16, info.signed, 8*ti.size, "0123456789abcdef", { .Prefix })
+                s = strconv.write_bits_128(buf[:], u, 16, info.signed, 8*uint(ti.size), "0123456789abcdef", { .Prefix })
 
             case:
-                s = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*ti.size, "0123456789", nil)
+                s = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*uint(ti.size), "0123456789", nil)
             }
         } else {
-            s = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*ti.size, "0123456789", nil)
+            s = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*uint(ti.size), "0123456789", nil)
         }
 
         _ = io.write_string(w, s) or_return
@@ -372,7 +372,7 @@ marshal_to_writer :: proc(w: io.Writer, v: any, opt: ^Marshal_Options) -> (err: 
                         case internal.Type_Info_Integer:
                             buf: [40]byte
                             u := cast_any_int_to_u128(ka)
-                            name = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*kti.size, "0123456789", nil)
+                            name = strconv.write_bits_128(buf[:], u, 10, info.signed, 8*uint(kti.size), "0123456789", nil)
                             
                             opt_write_key(w, opt, name) or_return
                         case: return .Unsupported_Type
